requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics==0.11.3
hydra-core>=1.3.1
